Ensure Build Bounding Box doesn't collapse
Previously the Bounding Box of a group of building was calculated
purely based on the position of each building.  Depending on the
geometry of the various buildings, this could result in a zero-volume
box, which would be culled away.

This change ensures that the BB is at least a 20x20x20 volume, by
increasing the convex hull size.

Fixes https://sourceforge.net/p/flightgear/codetickets/2542/