WS30: Increase loading priority for scenery tiles
Previously the database pager loaded PagedLOD based on distance
(once a PagedLoD is considered in-range).

As the same database pager is used for a variety of objects, this
meant that scenery tile loading was competing with STG files, AI
aircraft, buildings, random trees etc.

This change increases the priority of loading scenery tiles so tiles
should be inserted higher up the queue.