sound: Range test should not trigger 'once' sounds
This fixes the following behaviour.
Consider a sound definition of the form:

  <mode>once</mode>
  <property>trigger</property>
  <max-range>10</max-range>
  ...

Assume for instance that cockpit view is in range of the sound,
while external view is not. Then
(1) Set trigger=1, and keep it there -> sound plays (if in internal view)
(2) Switch to external view
(3) Switch back to internal view -> sound plays again!

Excepected behaviour is that the sound can only play immediately after
the  trigger goes to 1. If the range test did not exist, the sound would
not have played at step (3), so it should not play with the range test either.

This is because in SGXmlSound, the _active flag reflected both
the property/condition, and the out-of-range test.
Going out of range sets _active=false, which rearms the sound.

Change this so that
- the _active flag only depends on the property/condition
- independently, the out-of-range test stops and restarts the sample

This also affects the behaviour of dt_play and dt_stop. Going out
of range of the sample used to reset dt_play and increment dt_stop.
Now if the sample is active but out of range, dt_play will continue
incrementing.