SGPickAnimation: Deprecate duplicated knob/slider behaviour
Since commit 5c5e2a0f2f80 ("Remove debug output (and shorten test
names)") in 2013, knob and slider animations are installed multiple
times on objects which appear in the scene graph multiple times, such as
after an effect (material animation).

For some SGPickAnimations this is harmless, and we safely stop doing it
without breaking compatibility.

For others however, the motion would be reduced if the animation is only
installed once. Deprecate this behaviour for this case, and warn loudly
when it is detected, encouraging aircraft maintainers to work around the
issue by placing the animation node before the effect node in the XML.
This should ensure the aircraft works before and after the change is
eventually made.