WS30 - Save lots of memory on non-coastline tiles
Previously a coastline texture was created for tile irrespective
of whether any coastline was present for the tile.  This is was a
a huge waste of memory as the LoD scheme means that at the detailed
level most tiles do not contain coastline.

This change avoids creating coastline textures unless there is a
coastline present on the tile.  By observation this saves about
2GB of memory for areas where there is no coastline present.

Inspired by investigations by Poomklao T.  Many hands make like work.