Add SGSourceLocation based on sg_location
Add a new SGSourceLocation class, based on sg_location which is intended
for exceptions, but storing file paths as std::shared_ptr<std::string>,
rather than a 1Kb character array, with a static map of paths to shared
pointers to allow for automatic deduplication, and with exceptions
permitted.

Storing in a long array is fine for safely returning the path with
exceptions, but it makes it slightly cumbersome for use in linking
property nodes and derived scenegraph nodes back to the source XML files
for later debug output.