SGPickAnimation: Improve duplication warning debug output
Improve the debug output of knob/slider duplication warnings in several
ways:

- They now specify exact XML source locations of relevant nodes (see
  below).

- False positives due to duplicated object names between distinct
  objects is detected and trigger a separate warning.

- Only the second installation of an animation triggers the warnings
  (for cases where multiple duplicating animations precede the
  knob/slider animation and triggered multiple identical warnings).

- The expected multiplication of the animation is logged based on the
  number of ancestors with multiple parents.

- The blurb on first warning is now vaguer about the specific types of
  animation which trigger the issue (since it can be triggered by any
  animation which makes nodes appear multiple times in the scene graph).

To find the location of the knob/slider animation, the location stored
in the animation's configuration property node is logged.

To find the location of the node triggering the duplication, the scene
graph is searched upwards for nodes with multiple parents, and the first
valid parent location of each such node is logged. If there is no usable
location, the names of the node and its parents is logged to help
pinpoint what scene group needs a source location to be set.

In order to count how many times each object has had the animation
installed, they are now tracked using a std::string->unsigned int map,
instead of a duplicate string vector of object names remaining.