props: Store source locations in a separate mapping
The additional 24 bytes of SGSourceLocation in SGPropertyNode adds up to
10s or 100s of MiB for the sake of a very small fraction of
SGPropertyNode objects which actually use it.

To cut this additional memory usage down to just a few MiB, store this
data in a separate std::map (keyed by const SGPropertyNode*) outside of
the SGPropertyNode objects.

To avoid destructor ordering issues, where the map is destroyed prior to
the last SGPropertyNode objects, the map is allocated on demand and
destroyed once empty. A std::shared_mutex protects access to this global
data structure.

The only API change is that SGPropertyNode::getLocation now returns an
rvalue rather than a reference. This is for thread safety.