props_io: Set location on non-leaf nodes
Only set the SGPropertyNode original source location on non-leaf nodes,
This will reduce the amount of locations stored significantly without
impacting the usefulness, and will be particularly useful once the
location information is moved outside of SGPropertyNode.

Unfortunately this slightly complicates the line number tracking, since
a node isn't known to be non-leaf until PropsVisitor::endElement(), so
the line number at PropsVisitor::startElement() must be stored in the
State structure.