Explicitly set center of STG file LOD
Previously the LOD for a DelayLoadFileCallback was created
with a BoundingBox defined by the children of the LOD node.

This is fine for WS2.0 as by definition there will be some
terrain loaded, conveniently centered at the center of the tile.

For WS3.0 this is only the case if there are airports or other
OBJECT within the tile.  So a tile containing only BUILDING_LIST
or OBJECT_STATIC has no center at this point.  So OSG has it
at (0,0,0) or the center of the earth.

This means that the LOD and DelayLoadFileCallback never gets
triggered.

This change explicitly sets the center of the LOD as the center
of the tile.  This should result in no change for WS2.0 tile,
but means many more STG files are actually loaded in WS3.0.

Rick and Stuart suspect that this has been seen continually
during osm2city and WS3.0 development but previously assumed
just to be due to long loading times.