Use different vertex attribute for orthoscenery
Due to an OSG/FG bug, using vertex attribute aliasing and vertex
attributes with an index larger than 12 results in a crash. This commit
does not fix the issue entirely but it prevents crashes when attempting
to load the HDR pipeline on minimal settings.

It also makes more sense to use a dedicated texcoord attribute for
orthoscenery instead of a custom index. This does not matter too much in
modern OpenGL versions because it is our responsibility to assign vertex
attributes as we like.