WS30: Force sea to sea-level
Previously the elevation mesh was created purely based on the HGT,
and at runtime constrained to be underneath airport meshes.

This change also considers the landclass raster and forces all
vertices that are determined to be in the sea to 0 elevation,
removing some artifacts where areas of "sea" could be part way
up slopes.

A side effect of this is to create more attractive sea-cliffs.