WS30: Add ALERT for incompatible Atlas textures
When creating a texture atlas all the consistent images must
have the same size and internal format.  OSG provides a rather
useless warning message but only at rendering time.

This adds some explicit warning to simgear when the atlas is
created, including the affected filename and the internal
format code.  Unfortunately this is an integer that needs
to be looked up in /usr/include/GL/glext.h