Revert "props: Over-protect nodeOriginsLock"
This reverts commit 976b1372e631951752d391fc6fbd554611132eef. It was
dumb and a simpler solution is to just put the mutex on the heap and
never delete it.

props: Never destroy nodeOriginsLock

Allocate nodeOriginsLock on the heap without ever deleting it. This
ensures that it will never be destroyed, and should allow it to remain
usable even after static globals have been destroyed.

Suggested-by: Julian Smith <jules@op59.net>