WS30: BVH integration for ground materials
Previously the BVH tree for WS30 had no Material information
as Materials are not intrinsic to the terrain mesh (unlike WS20).

This change includes indice information in the BVH triangles, and
a lookup in the BVHTerrainTile to determine the material properties
of any intersection with the WS30 terrain by checking against the
landclass raster.