Fix Canvas disappearing when a window containing it is resized
This issue appeared when Canvas cameras became slave cameras instead
of nested cameras. The exact reason behind it is probably hidden deep
inside OSG or Canvas' lazy rendering.

A better approach to handling resizing in Canvas is to actually resize
the underlying texture and camera viewport, instead of recreating both
objects from scratch. This has the advantage of automagically fixing
the issue as well.