EffectCullVisitor: Fix copy constructor
The EffectCullVisitor copy constructor doesn't copy any of the
EffectCullVisitor members, so when cloned to the left and right cull
visitors of the SceneView and used for stereoscopic rendering, the
effect scheme is lost. This results in the wrong techniques being used
for rendering geometry to the gbuffer in the HDR compositor, which fails
due to the use of old style ALS shaders (GLSL version < 130) with vertex
attribute aliasing.

Copy both _collectLights and _effScheme so that the correct techniques
are chosen in this case.

Note this doesn't fully fix stereoscopic or VR rendering with the HDR
compositor, since only a subset of HDR compositor passes should be
treated as stereoscopic.