Fix network broadcast
Add new subclass of SGSocket called SGBroadcastSocket
Remove inappropriate insistance that 255.255.255.255 is the only possible broadcast address.
Remove the getBroadcast() predicate. Code in SGSocket is now only doing output to a
connected socket with send() and the broadcast socket option not enabled. Code in
SGBroadcastSocket is always doing output to an IP address with sendto() with the
broadcast socket option enabled.