Remove top-level .dirhash before running testUpdateNoChanges
During the first test (testBasicClone), some files inside dirB are
deleted or modified. This causes rebuild of dirB/.dirindex and
dirB/.dirhash, but does not update the top-level .dirhash.

That hash cache is only updated when the size or the last modified
time of dirB/.dirindex differs from the size or timestamp in the
.dirhash file. However, the size does not change in our case, and
the timestamp has granularity of one second, so if the test runs
quickly enough, the timestamps will match even after the contents of
dirB/.dirindex changed.

So testUpdateNoChanges was flaky: sometimes it passed, and sometimes
failed with "Bad request count" exception when comparing request
count for dirB. This change stabilizes it.