Compositor: Add support for compute passes
Add the ability for effects to specify compute-shaders, and add a new
compute pass type which dispatches a compute operation in order to use
them. This requires OpenGL 4.3 for it to be used.

The pass must specify the workgroup size (also specified in the compute
shader), and the global size which is used to calculate the workgroup
count. The global size can depend on the compositor viewport size (the
"screen" size), in which case it will be automatically resized when the
compositor is resized.