Compositor: Fix compute shaders broken resize
Compute shaders don't use framebuffer attachments, so OSG treats them as
using the main frame buffer. It therefore changes the camera's viewport
on window resize, which is shared with the compositor. This isn't
correct if the compositor is for off-screen rendering (e.g. for VR) and
results in visual breakage on resize in these cases.

This is fixed in a slightly hacky way in PassBuilder::build() by
matching the pass type to detect compute passes, and ensuring the
viewport is cloned so that resizes don't affect other compositor passes.